# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "audio-minutes API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/invitations": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "招待一覧",
        "operationId": "list_invitations_v1_admin_invitations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Invitations V1 Admin Invitations Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "招待発行 (URL は一度だけ返す)",
        "operationId": "create_invitation_v1_admin_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Invitation V1 Admin Invitations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/invitations/{invitation_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "招待取消",
        "operationId": "revoke_invitation_v1_admin_invitations__invitation_id__delete",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/users/{user_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "権限変更・無効化 (最後の owner は保護)",
        "operationId": "patch_user_v1_admin_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch User V1 Admin Users  User Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/retention": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "保持期間",
        "operationId": "get_retention_v1_admin_retention_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Retention V1 Admin Retention Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "保持期間の変更",
        "operationId": "put_retention_v1_admin_retention_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Put Retention V1 Admin Retention Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "監査ログ",
        "operationId": "get_audit_v1_admin_audit_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Audit V1 Admin Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/claude/login/{auth_session_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "ログイン状態 (URL は要求 owner にだけ返し保存しない)",
        "operationId": "claude_login_state_v1_admin_claude_login__auth_session_id__get",
        "parameters": [
          {
            "name": "auth_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Auth Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeLoginStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/claude/login/{auth_session_id}/code": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Claude認可codeの中継 (owner、再認証必須)",
        "operationId": "claude_login_code_v1_admin_claude_login__auth_session_id__code_post",
        "parameters": [
          {
            "name": "auth_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Auth Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaudeLoginCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeLoginActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/claude/login/{auth_session_id}/cancel": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "ログイン取消",
        "operationId": "claude_login_cancel_v1_admin_claude_login__auth_session_id__cancel_post",
        "parameters": [
          {
            "name": "auth_session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Auth Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeLoginActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClaudeLoginActionResponse": {
        "properties": {
          "auth_session_id": {
            "type": "string",
            "title": "Auth Session Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "url_ready",
              "completed",
              "failed",
              "cancelled",
              "expired"
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "auth_session_id",
          "state"
        ],
        "title": "ClaudeLoginActionResponse"
      },
      "ClaudeLoginCodeRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ClaudeLoginCodeRequest"
      },
      "ClaudeLoginStateResponse": {
        "properties": {
          "auth_session_id": {
            "type": "string",
            "title": "Auth Session Id"
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "url_ready",
              "completed",
              "failed",
              "cancelled",
              "expired"
            ],
            "title": "State"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          }
        },
        "type": "object",
        "required": [
          "auth_session_id",
          "state"
        ],
        "title": "ClaudeLoginStateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvitationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ],
            "title": "Role",
            "default": "member"
          },
          "expires_hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Expires Hours",
            "default": 24
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "InvitationRequest"
      },
...
      "RetentionSettings": {
        "properties": {
          "upload_hours": {
            "type": "integer",
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Upload Hours"
          },
          "audio_days": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Audio Days"
          },
          "log_days": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Log Days"
          }
        },
        "type": "object",
        "required": [
          "upload_hours",
          "audio_days",
          "log_days"
        ],
        "title": "RetentionSettings"
      },
...
      "UserPatch": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "owner",
                  "member"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "disabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled"
          }
        },
        "type": "object",
        "title": "UserPatch"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}